Move token allocators to separate file
